change log_service name to log
