+++ lib/common/components/wake_lock_and_brightness_enhancer.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:screen_brightness/screen_brightness.dart';
 import 'package:wakelock/wakelock.dart';
 
-import 'package:encointer_wallet/service/log/log_service.dart';
+import 'package:encointer_wallet/service/log/log.dart';
 
 /// Simple widget that sets the brightness and keeps the screen awake as long as the widget is alive.
 ///
