+++ lib/mocks/substrate_api/mock_assets_api.dart
@@ -1,6 +1,6 @@
 import 'package:encointer_wallet/mocks/data/mock_account_data.dart';
 import 'package:encointer_wallet/mocks/substrate_api/mock_js_api.dart';
-import 'package:encointer_wallet/service/log/log_service.dart';
+import 'package:encointer_wallet/service/log/log.dart';
 import 'package:encointer_wallet/service/substrate_api/assets_api.dart';
 
 class MockAssetsApi extends AssetsApi {
