+++ lib/mocks/substrate_api/mock_chain_api.dart
@@ -1,5 +1,5 @@
 import 'package:encointer_wallet/mocks/substrate_api/mock_js_api.dart';
-import 'package:encointer_wallet/service/log/log_service.dart';
+import 'package:encointer_wallet/service/log/log.dart';
 import 'package:encointer_wallet/service/substrate_api/chain_api.dart';
 
 class MockChainApi extends ChainApi {
