+++ lib/mocks/substrate_api/mock_encointer_api.dart
@@ -6,7 +6,7 @@ import 'package:encointer_wallet/models/ceremonies/ceremonies.dart';
 import 'package:encointer_wallet/models/communities/cid_name.dart';
 import 'package:encointer_wallet/models/communities/community_identifier.dart';
 import 'package:encointer_wallet/models/encointer_balance_data/balance_entry.dart';
-import 'package:encointer_wallet/service/log/log_service.dart';
+import 'package:encointer_wallet/service/log/log.dart';
 import 'package:encointer_wallet/service/substrate_api/encointer/encointer_api.dart';
 
 /// The key rationale behind this mock is that all the getters do not alter the app state.
