+++ lib/page-encointer/ceremony_box/ceremony_box.dart
@@ -14,7 +14,7 @@ import 'package:encointer_wallet/page-encointer/ceremony_box/meetup_info/compone
 import 'package:encointer_wallet/page-encointer/ceremony_box/meetup_info/meetup_info.dart';
 import 'package:encointer_wallet/page-encointer/common/encointer_map.dart';
 import 'package:encointer_wallet/page-encointer/meetup/ceremony_step1_count.dart';
-import 'package:encointer_wallet/service/log/log_service.dart';
+import 'package:encointer_wallet/service/log/log.dart';
 import 'package:encointer_wallet/service/substrate_api/api.dart';
 import 'package:encointer_wallet/service/tx/lib/tx.dart';
 import 'package:encointer_wallet/store/app.dart';
