+++ lib/page/profile/about_page.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 
 import 'package:encointer_wallet/common/components/jump_to_browser_link.dart';
-import 'package:encointer_wallet/service/log/log_service.dart';
+import 'package:encointer_wallet/service/log/log.dart';
 import 'package:encointer_wallet/utils/translations/index.dart';
 
 class AboutPage extends StatelessWidget {
