+++ lib/service/encointer_feed/feed.dart
@@ -6,7 +6,7 @@ import 'package:http/http.dart' as http;
 import 'package:encointer_wallet/config/consts.dart';
 import 'package:encointer_wallet/models/communities/community_identifier.dart';
 import 'package:encointer_wallet/models/index.dart';
-import 'package:encointer_wallet/service/log/log_service.dart';
+import 'package:encointer_wallet/service/log/log.dart';
 
 Future<List<MeetupOverrides>> getMeetupOverrides() async {
   final response = await http.get(Uri.parse(encointerFeedOverrides));
