+++ lib/service/log/log.dart
@@ -1,0 +1,15 @@
+import 'dart:developer';
+
+class Log {
+  static void e(String message, [String? description, StackTrace? stackTrace]) {
+    log('[ERROR] ${description ?? ''} ==> : $message ${stackTrace ?? ''}');
+  }
+
+  static void d(String message, [String? description, StackTrace? stackTrace]) {
+    log('[DEBUG] ${description ?? ''} ==> : $message ${stackTrace ?? ''}');
+  }
+
+  static void p(String message, [String? description, StackTrace? stackTrace]) {
+    log('[PRINT] ${description ?? ''} ==> : $message ${stackTrace ?? ''}');
+  }
+}
