+++ lib/service/meetup/feed_repo.dart
@@ -4,7 +4,7 @@ import 'package:http/http.dart' as http;
 
 import 'package:encointer_wallet/config/consts.dart';
 import 'package:encointer_wallet/service/meetup/feed_model.dart';
-import 'package:encointer_wallet/service/log/log_service.dart';
+import 'package:encointer_wallet/service/log/log.dart';
 
 class FeedRepo {
   FeedRepo([http.Client? client]) : _client = client ?? http.Client();
