+++ lib/service/substrate_api/core/dart_api.dart
@@ -2,7 +2,7 @@ import 'package:json_rpc_2/json_rpc_2.dart';
 import 'package:web_socket_channel/web_socket_channel.dart';
 
 import 'package:encointer_wallet/models/index.dart';
-import 'package:encointer_wallet/service/log/log_service.dart';
+import 'package:encointer_wallet/service/log/log.dart';
 
 /// Api to talk to an substrate node via the websocket protocol.
 ///
