+++ lib/service/substrate_api/encointer/encointer_dart_api.dart
@@ -1,7 +1,7 @@
 import 'package:encointer_wallet/models/communities/community_identifier.dart';
 import 'package:encointer_wallet/models/encointer_balance_data/balance_entry.dart';
 import 'package:encointer_wallet/models/index.dart';
-import 'package:encointer_wallet/service/log/log_service.dart';
+import 'package:encointer_wallet/service/log/log.dart';
 import 'package:encointer_wallet/service/substrate_api/core/dart_api.dart';
 
 class EncointerDartApi {
