+++ lib/store/data_update/data_update.dart
@@ -1,6 +1,6 @@
 import 'package:mobx/mobx.dart';
 
-import 'package:encointer_wallet/service/log/log_service.dart';
+import 'package:encointer_wallet/service/log/log.dart';
 
 part 'data_update.g.dart';
