+++ lib/utils/screenshot.dart
@@ -5,7 +5,7 @@ import 'dart:io';
 
 import 'package:flutter_driver/flutter_driver.dart';
 
-import 'package:encointer_wallet/service/log/log_service.dart';
+import 'package:encointer_wallet/service/log/log.dart';
 
 /// Called by integration test to capture images on the currently running device
 Future screenshot(FlutterDriver driver, Config config, String name,
